+++ tests/EscaperTest.php
@@ -6,6 +6,7 @@ namespace Duon\Boiler\Tests;
 
 use Duon\Boiler\Contract;
 use Duon\Boiler\Escaper;
+use Duon\Boiler\Exception\UnexpectedValueException;
 
 final class EscaperTest extends TestCase
 {
@@ -24,13 +25,27 @@ final class EscaperTest extends TestCase
 		$this->assertInstanceOf(Contract\Escaper::class, new Escaper());
 	}
 
-	public function testCanOverrideFlagsAndEncoding(): void
+	public function testCanUseHtmlStrategyExplicitly(): void
 	{
 		$escaper = new Escaper();
 
 		$this->assertSame(
-			'"quoted" &amp; &lt;tag&gt;',
-			$escaper->escape('"quoted" & <tag>', ENT_NOQUOTES),
+			'&lt;b&gt;&quot;Boiler&quot; &amp; more&lt;/b&gt;',
+			$escaper->escape('<b>"Boiler" & more</b>', Escaper::HTML),
 		);
+	}
+
+	public function testRejectsUnknownStrategy(): void
+	{
+		$this->throws(UnexpectedValueException::class, 'Unknown escape strategy `xml`');
+
+		new Escaper()->escape('<b>Boiler</b>', 'xml');
+	}
+
+	public function testRejectsUnknownDefaultStrategy(): void
+	{
+		$this->throws(UnexpectedValueException::class, 'Unknown escape strategy `xml`');
+
+		new Escaper('xml');
 	}
 }