+++ tests/WrapperTest.php
@@ -119,10 +119,9 @@ final class WrapperTest extends TestCase
 		$wrapper = new Wrapper(new class implements Escaper {
 			public function escape(
 				string $value,
-				int $flags = ENT_QUOTES | ENT_SUBSTITUTE,
-				string $encoding = 'UTF-8',
+				?string $strategy = null,
 			): string {
-				return strtoupper(htmlspecialchars($value, $flags, $encoding));
+				return strtoupper(htmlspecialchars($value));
 			}
 		});